refactor: backend code and functions
