+++ lib/dau_web/controllers/assessment_report_metadata_controller.ex
@@ -1,5 +1,5 @@
 defmodule DAUWeb.AssessmentReportMetadataController do
-  alias DAU.Feed.AssessmentReportMetadata
+  alias DAU.Feed.AssessmentReportMetadataRepository
   use DAUWeb, :controller
 
   @theme_radio_labels [
@@ -97,7 +97,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
     form_data = params["my_form"]
     form_data = Map.put(form_data, "feed_common_id", id)
 
-    case AssessmentReportMetadata.create(form_data) do
+    case AssessmentReportMetadataRepository.create(form_data) do
       {:ok, _metadata} ->
         conn
         |> put_flash(:info, "Assessment Report Metadata Added!")
@@ -116,7 +116,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
     id = params["id"]
 
     assessment_report_metadata =
-      AssessmentReportMetadata.get_assessment_report_metadata(id)
+      AssessmentReportMetadataRepository.get(id)
 
     changeset = Ecto.Changeset.change(assessment_report_metadata)
     form = Phoenix.HTML.FormData.to_form(changeset, as: :my_form)
@@ -127,7 +127,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
     id = params["id"]
     form_data = params["my_form"]
 
-    case AssessmentReportMetadata.update(id, form_data) do
+    case AssessmentReportMetadataRepository.update(id, form_data) do
       {:ok, _metadata} ->
         conn
         |> put_flash(:info, "Assessment Report Metadata Updated.")
@@ -143,7 +143,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
   def delete(conn, params) do
     id = params["id"]
 
-    case AssessmentReportMetadata.delete(id) do
+    case AssessmentReportMetadataRepository.delete(id) do
       {:ok, _metadata} ->
         conn
         # |> put_flash(:info, "assessment report metadata deleted successfully.")
@@ -157,7 +157,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
   end
 
   def fetch(conn, _params) do
-    metadata = AssessmentReportMetadata.fetch_all()
+    metadata = AssessmentReportMetadataRepository.fetch_all()
 
     translate_gender = fn gender_list ->
       gender_list
@@ -193,7 +193,7 @@ defmodule DAUWeb.AssessmentReportMetadataController do
         }
         |> Map.put_new(
           :assessment_report_url,
-          AssessmentReportMetadata.get_assessment_report_url(data.feed_common_id) || ""
+          AssessmentReportMetadataRepository.get_assessment_report_url(data.feed_common_id) || ""
         )
       end)
