+++ lib/dau_web/live/search_live/detail.html.heex
@@ -238,15 +238,17 @@
   </.section>
   <.section_assessment_report label="Assessment Report">
     <:form_button>
-      <div :if={Permission.has_privilege?(@current_user, :edit, Feed.AssessmentReportMetadata)}>
-        <div :if={Feed.AssessmentReportMetadata.report_exists?(@query.id) == false}>
+      <div :if={
+        Permission.has_privilege?(@current_user, :edit, Feed.AssessmentReportMetadataRepository)
+      }>
+        <div :if={Feed.AssessmentReportMetadataRepository.report_exists?(@query.id) == false}>
           <a href={"/demo/query/#{@query.id}/assessment-report/metadata"}>
             <button class="border border-green-200 bg-green-100 rounded-md hover:bg-green-200 hover:cursor-pointer px-2 py-1 w-fit">
               Create Form
             </button>
           </a>
         </div>
-        <div :if={Feed.AssessmentReportMetadata.report_exists?(@query.id) == true}>
+        <div :if={Feed.AssessmentReportMetadataRepository.report_exists?(@query.id) == true}>
           <a href={"/demo/query/#{@query.id}/assessment-report/metadata/edit"}>
             <button class="border border-yellow-200 bg-yellow-100 rounded-md hover:bg-yellow-200 hover:cursor-pointer px-2 py-1 w-fit">
               Edit Form
